==> 02 - Exportar tuits a dataframe.ipynb <==
from IPython.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))
from pymongo import MongoClient
from datetime import datetime
import pandas as pd

client = MongoClient()
db = client["twits_db"]
db.list_collection_names()
col = db["db.5e8fc63ddd8efca7e56c3215"]

- Filtrar por fechas
- Extraer columnas de interés
- RTs: extraer original si no fue procesado antes

start_ts = datetime.fromisoformat('2020-04-12T00:00:00+00:00')
end_ts =  datetime.fromisoformat('2020-04-19T00:00:00+00:00')

# Convert string to datetime in the aggregation pipeline and apply projection
pipeline = [        
    {
        '$project': {
            "id": 1,
            "user.id": 1,
            "text": 1,
            "truncated": 1,
            'extended_tweet.full_text': 1,
            "in_reply_to_status_id": 1,
            "in_reply_to_user_id": 1,
            "_id": 0,
            'created_at': {
                '$dateFromString': {
                    'dateString': '$created_at'
                }
            },
            "retweeted_status.id": 1,
            "retweeted_status.user.id": 1,
            "retweeted_status.text": 1,
            "retweeted_status.extended_tweet.full_text": 1,
        }
    },
    {
        "$match": {
            'created_at': {
                '$gte': start_ts,
                '$lte': end_ts
            },
        }
    }
]

# Batch settings
batch_size = 1000000  # Define the size of each batch
batch_number = 1  # Counter for batch files

# Batch processing
batch = []

def save_batch(batch, batch_number):
    # Convert the batch to a DataFrame and save as CSV (or use csv.writer for raw CSV)
    df = pd.DataFrame(batch)
    for f in ["reply_to_tweet_id", "reply_to_user_id", "retweet_tweet_id", "retweet_user_id"]:
        df[f] = df[f].astype("Int64")        
    df.to_csv(f'data/tweets_batch_{batch_number:03d}.tsv', index=False, sep='\t')
    print(f"Saved batch {batch_number:03d}")

for d in col.aggregate(pipeline):
    row = {
        "tweet_id": d["id"],
        "user_id": d["user"]["id"],        
        "created_at": d["created_at"],
        "full_text": d["text"],
        "reply_to_tweet_id": d["in_reply_to_status_id"],
        "reply_to_user_id": d["in_reply_to_user_id"],
    }
    if "extended_tweet" in d:
        row["full_text"] = d["extended_tweet"]["full_text"]
    if "retweeted_status" in d:
        dr = d["retweeted_status"]
        row |= {
            "retweet_tweet_id": dr["id"],
            "retweet_user_id": dr["user"]["id"],            
        }
        if "extended_tweet" in dr:
            row["full_text"] = dr["extended_tweet"]["full_text"]
        else:
            row["full_text"] = dr["text"]

    batch.append(row)
    # Save the batch to CSV when it reaches the batch_size
    if len(batch) == batch_size:
        save_batch(batch, batch_number)
        
        # Reset the batch and increment the batch number
        batch = []
        batch_number += 1

# Save the final batch if it's not empty
if batch:
    save_batch(batch, batch_number)

print(f"Processed {batch_number} batches.")
    

chunk_size = 10000  # Adjust the chunk size as needed
all_chunks = []

for i in range(1,8):
    chunks = pd.read_csv(f'data/tweets_batch_00{i}.tsv', sep='\t',
                         engine='python',
                         on_bad_lines='skip',
                         chunksize=chunk_size)
    all_chunks += chunks

df = pd.concat(all_chunks, ignore_index=True)

for f in ["user_id", "reply_to_tweet_id", "reply_to_user_id", "retweet_tweet_id", "retweet_user_id"]:
    df[f] = df[f].astype("Int64")

df

### How many replies?

(~df.reply_to_tweet_id.isna()).mean()

(~df.reply_to_tweet_id.isna()).sum()

replied_in_df = df.reply_to_tweet_id.isin(df.tweet_id)

replied_in_df.mean()

replied_in_df.sum()

dfr = df[df.tweet_id.isin(df[replied_in_df].tweet_id)]
# Exclude self-replies (threads)
dfr = dfr[dfr.user_id != dfr.reply_to_user_id]

dfr = pd.merge(left=dfr.sample(10),
               right=df,
               left_on="reply_to_tweet_id", right_on="tweet_id",
               how='left')

dfr

for _, r in dfr.iterrows():
    print("=================")
    print(r.full_text_y)
    print("-----------------")
    print(r.full_text_x)

pd.merge?

### retweets

df.retweet_tweet_id.isna().mean()

rt_counts = df.reply_to_tweet_id.value_counts()

import plotly.express as px
px.bar(rt_counts.value_counts())
