==> 03.1 - embeddings.ipynb <==
Limitamos el análisis al día más activo de la semana.
Sólo para este día traemos embeddings y hacemos topic modeling.

Posiblemente el topic modeling requiera tomar una muestra de los tuits, para luego extenderlo al día completo.m

import numpy as np
import pandas as pd
import random

chunk_size = 10000  # Adjust the chunk size as needed
all_chunks = []

for i in range(1,8):
    chunks = pd.read_csv(f'data/tweets_batch_00{i}.tsv', sep='\t',
                         usecols=["tweet_id", "user_id", "full_text", "created_at"],
                         engine='python',
                         on_bad_lines='skip',
                         chunksize=chunk_size)
    all_chunks += chunks

df = pd.concat(all_chunks, ignore_index=True)

for f in ["user_id"]:
    df[f] = df[f].astype("Int64")

df_full = df

df_full.columns

df_full.shape

df = df_full

df_full.created_at = pd.to_datetime(df_full.created_at, errors='coerce')
df_full.created_at.min(), df_full.created_at.max()

df_full["created_date"] = df_full.created_at.dt.date
day_counts = df_full.created_date.value_counts()
day_counts

# Tomamos el día más activo
day_max = day_counts.idxmax()

df_max_day = df_full[df_full.created_date == day_max]

len(df_max_day)

df_max_day.to_csv(f"data/covid_twitter_sample_{day_max.strftime('%Y-%m-%d')}.csv", index=False)

docs = df_max_day.full_text.values.tolist()

docs[:2]

docs[0]

from sentence_transformers import SentenceTransfo
rmer

# Spanish multilingual model (lightweight)
# This takes ages!
# model = SentenceTransformer("distiluse-base-multilingual-cased-v1")
embeddings_model = SentenceTransformer("paraphrase-multilingual-MiniLM-L12-v2")

embeddings_model

embeddings = embeddings_model.encode(docs, batch_size=1024, show_progress_bar=True)

import numpy as np

np.save(f"data/covid_twitter_sample_{day_max.strftime('%Y-%m-%d')}.npy", embeddings)
