==> 03.2 - topic model train on sample.ipynb <==
## 1. Cargar tuits

import numpy as np
import pandas as pd
import random

%ls data/

df = pd.read_csv("data/covid_twitter_sample_2020-04-16.csv")

df.columns

df.shape

df_full = df

df_full.shape

import numpy as np

embeddings = np.load(f"data/covid_twitter_sample_2020-04-16.npy")

embeddings.shape

docs = df.full_text.dropna().values.tolist()

len(docs)

docs[:2]

docs[0]

3. Topic modeling

## Hashtags populares

import re

import unicodedata

def remove_accents(input_str):
    nfkd_form = unicodedata.normalize('NFKD', input_str)
    only_ascii = nfkd_form.encode('ASCII', 'ignore')
    return str(only_ascii)

def extract_hashtags(s):
    return [
        remove_accents(ht.lower())
        for ht in re.findall(r"#(\w+)", s)
    ]

df.head()

df["full_text"] = df.full_text.astype(str)

df["hashtags"] = df.full_text.apply(extract_hashtags)

df.head()

from collections import Counter

ht_counts = Counter()

for hts in df.hashtags.values:
    ht_counts.update(hts)

ht_counts.most_common(100)

len(df)

# BERTopic

from sentence_transformers import SentenceTransformer

embedding_model = "paraphrase-multilingual-MiniLM-L12-v2"

# Spanish multilingual model (lightweight)
model = SentenceTransformer(embedding_model)

model

from bertopic import BERTopic
from bertopic.representation import KeyBERTInspired

# Create your representation model
representation_model = KeyBERTInspired()

sample_rate = 1.0

sample_inds = np.random.choice(len(docs), int(sample_rate * len(docs)), replace=False)

sample_inds.dtype

type(docs)

docs_sample = [docs[i] for i in sample_inds]

len(docs_sample)

embeddings_sample = embeddings[sample_inds]

embeddings_sample.shape

import spacy
import string

# Load the Spanish library from SpaCy
nlp = spacy.load("es_core_news_sm")

# Create list of punctuation marks
punctuations = string.punctuation

# Create list of stopwords from spaCy
es_stopwords = spacy.lang.es.stop_words.STOP_WORDS

es_stopwords = list(es_stopwords) + ["https", "http", "com", "covid", "covid19", "19", "co", "coronavirus", "rt"]

# --- 4. Vectorizer for topic words (Spanish stopwords + unigrams/bigrams)
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction import text

vectorizer_model = CountVectorizer(
    stop_words=es_stopwords,
    ngram_range=(1, 2),        # include bigrams to catch phrases like "cambio climático"
    min_df=5                   # ignore rare words
)

# Create and fit the BERTopic model using the precomputed embeddings
topic_model = BERTopic(
    language="spanish",
    vectorizer_model=vectorizer_model,
    verbose=True
)

# topics, _ = topic_model.fit_transform(docs, embeddings)
topics, _ = topic_model.fit_transform(docs_sample, embeddings_sample)

# View topic info
topic_info = topic_model.get_topic_info()

# Method 1 - safetensors
embedding_model = "sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2"
topic_model.save("covid_twitter_topic_model_2020-04-16-100p", serialization="safetensors", save_ctfidf=True, save_embedding_model=embedding_model)
