==> 03.3 - topic model extend to full.ipynb <==
## Cargar tuits y embeddings

import pandas as pd

df = pd.read_csv("data/covid_twitter_sample_2020-04-16.csv")
df.columns

df.shape

df_full = df

import numpy as np

embeddings = np.load(f"data/covid_twitter_sample_2020-04-16.npy")
embeddings.shape

docs = df.full_text.dropna().values.tolist()
len(docs)

docs[:2]

# Cargar topic model

from bertopic import BERTopic

topic_model = BERTopic.load("covid_twitter_topic_model_2020-04-16")

topic_model.get_topic_info().head(100)

# Aplicar topic model al conjunto completo de tuits

topics_full = topic_model.transform(docs)

# TODO: asignar tema principal a cada tuit

topics_full

[x.shape for x in topics_full]

# Guardar las asignaciones de temas del dataset completo

df_full['topic'] = topics_full[0]
df_full['topic_prob'] = topics_full[1]

df_full[['tweet_id', 'topic', 'topic_prob']].to_csv("data/covid_twitter_sample_2020-04-16_topics.csv", index=False)

df_full.topic.value_counts().head(20)
