==> 03.4 - topic model analyze topics.ipynb <==
# Cargar tuits y temas

import pandas as pd

df = pd.read_csv("data/covid_twitter_sample_2020-04-16.csv")
df.columns

df_temas = pd.read_csv("data/covid_twitter_sample_2020-04-16_topics.csv")
df_temas.columns

df_tuits_temas = df.merge(df_temas, on="tweet_id", how="inner")
df_tuits_temas.columns

df_topic_info = pd.read_csv("data/covid_twitter_sample_2020-04-16_topics-info.csv")
df_topic_info.columns

df_topic_info.shape

def visualize_topic(topic_ind, n_tuits_topic=20, print_output=True):
    # show topic name
    topic_title = df_topic_info[df_topic_info.Topic == topic_ind]['Name'].values[0]
    result_string = "\n=============================="
    result_string += f"\n===== Topic {topic_ind:04d}: {topic_title}"

    # collect tuits with highest score
    df_topic_tuits = df_tuits_temas[df_tuits_temas.topic == topic_ind].drop_duplicates("full_text").sort_values(by="topic_prob", ascending=False).head(n_tuits_topic)

    for t in df_topic_tuits.full_text.values:
        result_string += '\n' + t + "\n-----"

    if print_output:
        print(result_string)
    return result_string

visualize_topic(9)

visualize_topic(1)

visualize_topic(0)

visualize_topic(2)

visualize_topic(3)

visualize_topic(4)

visualize_topic(5)

visualize_topic(6)

visualize_topic(7)

visualize_topic(8)

visualize_topic(18)

# Calcular reply count/rate por tema
- Cantidad total de respuestas a tuits de cada tema
- Divididos sobre cantidad de tuits en el tema

chunk_size = 10000  # Adjust the chunk size as needed
all_chunks = []

for i in range(1,8):
    chunks = pd.read_csv(f'data/tweets_batch_00{i}.tsv', sep='\t',
                         usecols=["user_id", "tweet_id", "created_at",
                                  "reply_to_user_id", "reply_to_tweet_id", "full_text"],
                         engine='python',
                         on_bad_lines='skip',
                         chunksize=chunk_size)
    all_chunks += chunks

df_replies = pd.concat(all_chunks, ignore_index=True)

def get_reply_count(topic_ind):
    # get tuits for topic
    df_topic_tuits = df_tuits_temas[df_tuits_temas.topic == topic_ind]

    # get replies for tuits in topic
    df_replies_topic = df_replies[df_replies.reply_to_tweet_id.isin(df_topic_tuits.tweet_id.values)]

    # excluimos autorespuestas (hilos)
    df_replies_topic = df_replies_topic[df_replies_topic.user_id != df_replies_topic.reply_to_user_id]

    # count
    n_replies = df_replies_topic.shape[0]
    return n_replies

df_topic_info["reply_count"] = df_topic_info.Topic.apply(get_reply_count)
df_topic_info["reply_rate"] = df_topic_info.reply_count / df_topic_info.Count

def get_active_and_engaging_topics(n=100):
    df_active = df_topic_info[["Topic", "Count"]].sort_values(by="Count", ascending=False).head(n)

    df_engaging = df_topic_info[["Topic", "reply_count"]].sort_values(by="reply_count", ascending=False).head(n)

    active_and_engaging_ids = set(df_active.Topic).intersection(df_engaging.Topic)

    df_active_and_engaging = df_topic_info[df_topic_info.Topic.isin(active_and_engaging_ids)]
    return df_active_and_engaging

df_active_and_engaging = get_active_and_engaging_topics(n=1000)
len(df_active_and_engaging)

df_active_and_engaging

# Visualizar replies en topics polémicos

def visualize_replies(topic_ind, n_replies=10, print_output=True):
    # get tuits for topic
    df_topic_tuits = df_tuits_temas[df_tuits_temas.topic == topic_ind]

    result_string = "\n-+-+-+-+-+-+-+-+-+-+-+-+-+-+"
    result_string += "\nTuits and replies for topic " + str(topic_ind) + ":"
    for tweet_id in df_topic_tuits.tweet_id.values:
        # get replies for tuits in topic
        df_replies_tuit = df_replies[df_replies.reply_to_tweet_id == tweet_id]
        if len(df_replies_tuit) > 0:
            if len(df_replies_tuit) > n_replies:
                df_replies_tuit = df_replies_tuit.sample(n_replies)
            result_string += "\nTuit:"
            tweet_text = df_topic_tuits[df_topic_tuits.tweet_id == tweet_id].full_text.values[0]
            result_string += '\n' + tweet_text
            result_string += "\nReplies:"

            for reply in df_replies_tuit.full_text.values:
                reply_line = "\n\t-" + reply.replace('\n', '\n\t')
                result_string += reply_line

            result_string += "-----"
    if print_output:
        print(result_string)
    return result_string

visualize_replies(0)

# Seleccionar temas
entre los más populares (nro de tuits) y más polémicos (reply rate)

topics_report = ""
for topic_ind in df_active_and_engaging.Topic.values:
    print(topic_ind)
    topics_report += visualize_topic(topic_ind, n_tuits_topic=5, print_output=False)
    topics_report += visualize_replies(topic_ind, n_replies=10, print_output=False)

with open("data/covid_twitter_sample_2020-04-16_topics_report.txt", "w", encoding="utf-8") as f:
    f.write(topics_report)

df_active_and_engaging

# Selección de tuits relevantes y polémicos
- En top-k de temas
- Con más de n respuestas de otros

def get_replies_tuit(tweet_id, user_id):
    # get replies for tuits in topic
    df_replies_tweet = df_replies[(df_replies.reply_to_tweet_id == tweet_id) & (df_replies.user_id != user_id)]

    return df_replies_tweet

def get_active_and_engaging_tuits(n_topics=1000, n_replies=10):
    df_active = df_topic_info[["Topic", "Count"]].sort_values(by="Count", ascending=False).head(n_topics)

    df_tuits_active = df_tuits_temas[df_tuits_temas.topic.isin(df_active.Topic.values)]

    # with at least one reply
    have_replies = df_replies.reply_to_tweet_id.unique()
    df_tuits_active = df_tuits_active[df_tuits_active.tweet_id.isin(have_replies)]

    df_tuits_active["reply_count"] = df_tuits_active.apply(lambda x: len(get_replies_tuit(x.tweet_id, x.user_id)), axis=1)

    df_tuits_active_and_engaging = df_tuits_active[df_tuits_active.reply_count >= n_replies]
    return df_tuits_active_and_engaging

active_and_engaging_tuits = get_active_and_engaging_tuits(n_topics=5000, n_replies=3)
active_and_engaging_tuits

active_and_engaging_tuits.reply_count.sum()

active_and_engaging_tuits.tweet_id.nunique()

## Guardamos interacciones para análisis posterior

df_replies.columns

df_interactions = pd.merge(
    df_replies.rename(columns={"full_text": "reply_text",
                               "tweet_id": "reply_tweet_id",
                               "user_id": "reply_user_id",
                               "created_at": "reply_created"}),
    active_and_engaging_tuits.rename(columns={"full_text": "original_tweet_text",
                                              "tweet_id": "original_tweet_id",
                                              "user_id": "original_user_id",
                                              "created_at": "original_created"}),
    how="inner",
    left_on="reply_to_tweet_id",
    right_on="original_tweet_id"
)[[
    'created_date', 'topic', 'topic_prob',
    'original_created', 'original_tweet_id', 'original_tweet_text', 'original_user_id',
    'reply_created', 'reply_tweet_id', 'reply_text', 'reply_user_id'
]].query('original_user_id != reply_user_id') # exclude self-replies

df_interactions

df_interactions.shape

df_interactions.to_csv("data/covid_twitter_sample_2020-04-16_interactions.csv", index=False)
